Defer numba guvectorize compilation for display functions
